pkgs/feature-template/requirements.txt
cookiecutter
jinja2-strcase
pyyaml